requirements.txt
asgiref==3.3.4
boto3==1.17.54
botocore==1.20.54
dj-database-url==0.5.0
Django==3.2
django-cors-headers==3.7.0
django-environ==0.4.5
django-filter==2.4.0
django-storages==1.11.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
gunicorn==20.1.0
jmespath==0.10.0
psycopg2-binary==2.8.6
pycountry==20.7.3
PyJWT==2.0.1
python-dateutil==2.8.1
pytz==2021.1
s3transfer==0.4.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
whitenoise==5.2.0
